requirements.txt
werkzeug
PyJWT
sqlalchemy
Flask~=1.1
gunicorn==20.1.0
